Setup grabbing values from a range of timeframes.

+++ api/src/main/kotlin/me/stegall/services/Values.kt
@@ -5,9 +5,9 @@ import me.stegall.data.Value
 
 class Values {
   fun toValue(position: Position, time: String): Value {
-    return Value(position.symbol, position.qty, position.current_price, time)
+    return Value(position.symbol, position.qty.toLong(), position.current_price.toFloat(), time)
   }
   fun toValues(positions: ArrayList<Position>, time: String): ArrayList<Value> {
-    return ArrayList<Value>(positions.map { Value(it.symbol, it.qty, it.current_price, time) })
+    return ArrayList<Value>(positions.map { Value(it.symbol, it.qty.toLong(), it.current_price.toFloat(), time) })
   }
 }
